+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 Change Log
 ==========
 
-Version 1.0.1 *(2012-10-28)*
+Version 1.1 *(2013-03-07)*
+--------------------------
+
+- New API version: 2 (Californication)
+  * [get|set]MarkerHeight renamed to [get|set]BottomOffset
+  * [is|set]Actionnable renamed to [is|set]Interactive
+  * New `OnMapViewClickListener` to listen to taps
+- Fix: Incorrect measurement of the `MapCalloutDrawable` intrinsic height
+
+Version 1.0.1 *(2012-10-23)*
 --------------------------
 
 - General code and Javadoc cleanup